// code-app/src/generated/models/job-application-model.ts
import type { BusinessGroup } from './business-group-model';
import type { Company } from './company-model';

export const JobApplicationArrangementKeyToLabel = {
  'Remote': 'Remote',
  'OnSite': 'On-site',
  'Hybrid': 'Hybrid'
} as const;
export type JobApplicationArrangementKey = keyof typeof JobApplicationArrangementKeyToLabel;

export const JobApplicationStageKeyToLabel = {
  'Researching': 'Researching',
  'Applied': 'Applied',
  'Interviewing': 'Interviewing',
  'Offer': 'Offer',
  'Closed': 'Closed'
} as const;
export type JobApplicationStageKey = keyof typeof JobApplicationStageKeyToLabel;

export interface JobApplication {
  /**
   * @displayName Job Application
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Role
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  role: string;
  /**
   * @displayName Arrangement
   */
  arrangementKey?: JobApplicationArrangementKey;
  /**
   * @displayName Business Group
   */
  businessGroup?: Pick<BusinessGroup, 'id' | 'businessGroupName'>;
  /**
   * @displayName City
   */
  city?: string;
  /**
   * @displayName Company
   * @validationRule Required for create/update operations
   */
  company: Pick<Company, 'id' | 'companyName'>;
  /**
   * @displayName Date Applied
   */
  dateApplied?: string;
  /**
   * @displayName Job ID
   */
  jobID?: string;
  /**
   * @displayName Job Link
   */
  jobLink?: string;
  /**
   * @displayName Next Step
   */
  nextStep?: string;
  /**
   * @displayName Notes
   */
  notes?: string;
  /**
   * @displayName Stage
   * @validationRule Required for create/update operations
   */
  stageKey: JobApplicationStageKey;
}

export const _JobApplication = 'JobApplication' as const;